+++ backend/src/routes/spa/mod.rs
@@ -1,10 +1,10 @@
 use actix_web::dev::HttpServiceFactory;
 
-mod spa_impl;
-
 #[cfg(feature = "spa")]
 pub fn spa() -> impl HttpServiceFactory {
-    spa_impl::Spa::default()
+    use actix_web_nextjs_spa::Spa;
+
+    Spa::default()
         .index_file("./web/index.html")
         // .static_resources_mount("./web/")
         .static_resources_location("./web")